requirements.txt
discord.py
toml
colored
pickledb